+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 
 - [beta] Fix an issue where long matches were not highlighted in folder search results.
+- [beta] Fix an issue where binary files, such as executables without a filename extension, were searched in the folder find.
 
 
 ### Known Issues